# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SettingsCommentPermission } from './settingsCommentPermission';

export interface Settings {
  discoverable: boolean;
  showAge: boolean;
  showRegion: boolean;
  notificationsEnabled: boolean;
  commentPermission: SettingsCommentPermission;
  deleteRequested: boolean;
}